requirements.txt
requests
minio
psycopg2-binary
